Adds search to wizard spellbook
